requirements.txt
PyMuPDF==1.24.9
python-docx==1.1.2
python-pptx==0.6.23